pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "pydantic-settings>=2.8.1",
    "sqlalchemy[asyncio]>=2.0.39",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]

[tool.ruff]
exclude = ["env.py", "volume"]

[tool.pytest.ini_options]
addopts = "--ignore=volume"
